Add support for communication with backend
