--- data.json
@@ -1,9 +1,0 @@
-{
-  "memory_limit": "250",
-  "time_limit": "0.5",
-  "compilation": "/usr/bin/g++ -o2 -static -pipe -o source source.cpp",
-  "execution": "./source < main.in > main.out",
-  "checker": "checker.cpp",
-  "work_directory": "/tmp/run",
-  "volumen": "/<absolute path>/judge-bot/run1"
-}
