+++ package.json
@@ -7,7 +7,7 @@
     "test": "test"
   },
   "scripts": {
-    "postinstall": "mkdir -p data && mkdir -p data/submissions && mkdir -p data/problems",
+    "postinstall": "mkdir -p data && mkdir -p data/submissions && mkdir -p data/problems && mkdir -p data/runs",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -26,6 +26,8 @@
   },
   "homepage": "https://github.com/in-silico/judge-bot#readme",
   "dependencies": {
-    "node-uuid": "^1.4.7"
+    "mkdirp": "^0.5.1",
+    "node-uuid": "^1.4.7",
+    "path": "^0.12.7"
   }
 }
